+++ acceptance/openstack/ims/CreateImageFromOBS_test.go
@@ -1,13 +1,9 @@
 package ims
 
 import (
-	"io"
-	"net/http"
-	"os"
 	"strings"
 	"testing"
 
-	golangsdk "github.com/opentelekomcloud/gophertelekomcloud"
 	"github.com/opentelekomcloud/gophertelekomcloud/acceptance/clients"
 	"github.com/opentelekomcloud/gophertelekomcloud/acceptance/tools"
 	"github.com/opentelekomcloud/gophertelekomcloud/openstack/ims/v1/others"
@@ -15,16 +11,6 @@ import (
 	"github.com/opentelekomcloud/gophertelekomcloud/openstack/obs"
 	th "github.com/opentelekomcloud/gophertelekomcloud/testhelper"
 )
-
-func getClient(t *testing.T) (*golangsdk.ServiceClient, *golangsdk.ServiceClient) {
-	v1, err := clients.NewIMSV1Client()
-	th.AssertNoErr(t, err)
-
-	v2, err := clients.NewIMSV2Client()
-	th.AssertNoErr(t, err)
-
-	return v1, v2
-}
 
 func TestCreateImageFromOBS(t *testing.T) {
 	client1, client2 := getClient(t)
@@ -43,24 +29,7 @@ func TestCreateImageFromOBS(t *testing.T) {
 	})
 	th.AssertNoErr(t, err)
 
-	img, err := os.CreateTemp("", "ims-rancher.img")
-	th.AssertNoErr(t, err)
-	t.Cleanup(func() {
-		err = img.Close()
-		th.AssertNoErr(t, err)
-		err = os.Remove(img.Name())
-		th.AssertNoErr(t, err)
-	})
-
-	resp, err := http.Get("https://releases.rancher.com/os/latest/rancheros-openstack.img")
-	th.AssertNoErr(t, err)
-	t.Cleanup(func() {
-		err = resp.Body.Close()
-		th.AssertNoErr(t, err)
-	})
-
-	_, err = io.Copy(img, resp.Body)
-	th.AssertNoErr(t, err)
+	img, err := downloadIMG(t)
 
 	objectName := tools.RandomString("ims-test-", 5)
